// frontend/src/openapi/models/FeedbackCreateCategoryEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `bug` - Report Bug
 * * `improvement` - Suggest Improvement
 * * `question` - Ask a Question
 * * `love` - Share What You Love
 * * `feature` - Request Feature
 * * `performance` - Performance Issue
 * * `accessibility` - Accessibility Issue
 * * `security` - Security Concern
 * * `documentation` - Documentation Issue
 * * `other` - Other
 * @export
 */
export const FeedbackCreateCategoryEnum = {
    bug: 'bug',
    improvement: 'improvement',
    question: 'question',
    love: 'love',
    feature: 'feature',
    performance: 'performance',
    accessibility: 'accessibility',
    security: 'security',
    documentation: 'documentation',
    other: 'other'
} as const;
export type FeedbackCreateCategoryEnum = typeof FeedbackCreateCategoryEnum[keyof typeof FeedbackCreateCategoryEnum];


export function instanceOfFeedbackCreateCategoryEnum(value: any): boolean {
    for (const key in FeedbackCreateCategoryEnum) {
        if (Object.prototype.hasOwnProperty.call(FeedbackCreateCategoryEnum, key)) {
            if (FeedbackCreateCategoryEnum[key as keyof typeof FeedbackCreateCategoryEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FeedbackCreateCategoryEnumFromJSON(json: any): FeedbackCreateCategoryEnum {
    return FeedbackCreateCategoryEnumFromJSONTyped(json, false);
}

export function FeedbackCreateCategoryEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): FeedbackCreateCategoryEnum {
    return json as FeedbackCreateCategoryEnum;
}

export function FeedbackCreateCategoryEnumToJSON(value?: FeedbackCreateCategoryEnum | null): any {
    return value as any;
}

export function FeedbackCreateCategoryEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): FeedbackCreateCategoryEnum {
    return value as FeedbackCreateCategoryEnum;
}